user password update, user update, user verified Scope

+++ app/Http/Controllers/Users/UserAccountController.php
@@ -6,9 +6,11 @@ use App\Http\Controllers\ApiController;
 use App\Models\Renter;
 use App\Models\User;
 use Carbon\Carbon;
+use Illuminate\Validation\Rule;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Validation\ValidationException;
@@ -19,12 +21,13 @@ class UserAccountController extends ApiController
 
     public function __construct()
     {
-        $this->middleware(['auth:users','verified:verification.verify'])->except('store','index','show');
+        $this->middleware(['auth:users','admin.auth'])->only('approve');
+        $this->middleware(['auth:users','verified:verification.verify'])->except('approve','store','index','show');
     }
 
     public function index(Request $request)
     {
-        $users = User::basic()->paginate(10);
+        $users = User::basic()->verified()->paginate(10);
         return $this->success($users);
     }
 
@@ -101,30 +104,48 @@ class UserAccountController extends ApiController
      * @return JsonResponse
      * @throws ValidationException
      */
-    public function update(Request $request, User $user)
+    public function update(Request $request, $userId)
     {
+        /* check if user exists */
+        $user = User::find($userId);
+        if(is_null($user)) {
+            return response()->json(['success'=> true, 'message' => 'no record found'], 403);
+        }
+
         $rules = [
             'fname' => 'required|min:3|max:30',
             'lname' => 'required|min:3|max:30',
-            'email' => 'required|email|unique:users,email,' . $user->id,
-            'password' => 'required|confirmed',
+            'email' => [
+                'required',
+                'email', 
+                Rule::unique('users')->ignore($user->id)
+            ],
+            'password' => 'required',
             'phone_number' => 'sometimes|required|regex:/^0[567]{1}[0-9]{8}$/i',
             'dob' => 'required|date',
             'profile_pic' => 'sometimes|required|image|mimes:jpg,png',
             'identity_pic' => 'sometimes|required|image|mimes:jpg,png',
         ];
-
+        /* validate */
         $validate = Validator::make($request->all(), $rules);
         if($validate->fails()) {
             return response()->json(['success' => false, 'errors' =>$validate->errors()], 403);
         }
+        /* check if the password match */
+        if(!Hash::check($request->password, $user->password)) {
+            return response()->json(['success' => false, 'message' =>"password does not match"], 401);
+        }
+        /* check if the user is authorized to update */
         $inspect = Gate::inspect('update', $user);
         if($inspect->denied()) {
             return response()->json(['success' => false, 'errors' =>$inspect->message()], 401);
         }
+        /* hash the password */
+        $validate = $validate->validated();
+        $validate['password'] = bcrypt($request->password);
+        $user->update($validate);
 
-        $user->update($validate->validated());
-
+        /* if there is a profile_pic in the request update it */
         if($request->hasFile('profile_pic')) {
             $user->profile_pic = str_replace("uploads/{$user->user_role}/","", $user->profile_pic);
             Storage::disk($user->user_role)->delete($user->profile_pic);
@@ -135,6 +156,7 @@ class UserAccountController extends ApiController
                         $user->user_role
                     );
         }
+        /* if there is a identity in the request update it */
         if($request->hasFile('identity_pic')) {
             $user->identity_pic = str_replace("uploads/{$user->user_role}/","",$user->identity_pic);
             Storage::disk($user->user_role)->delete($user->identity_pic);
@@ -146,10 +168,44 @@ class UserAccountController extends ApiController
                     );
         }
 
-
+        /* save all */
         $user->save();
         return response()->json(['success'=> true, 'message' => $user], 200);
     }
+    public function password_reset(Request $request,$userId) {
+        $user = User::find($userId);
+        if(is_null($user)) {
+            return response()->json(['success'=> true, 'message' => 'no record found'], 403);
+        }
+        $rules = [
+            'old_password' => 'required',
+            'password' => 'required|confirmed',
+        ];
+
+        $validate = Validator::make($request->all(), $rules);
+        if($validate->fails()) {
+            return response()->json(['success' => false, 'errors' =>$validate->errors()], 403);
+        }
+
+        if(!Hash::check($request->old_password, $user->password)) {
+            return response()->json(['success' => false, 'message' =>"password does not match"], 401);
+        }
+
+        $user->password = bcrypt($request->password);
+        $user->save();
+        
+        return response()->json(['success'=> true, 'message' => $user], 200);
+    }
+
+    public function approve($userId) {
+        $user = User::find($userId);
+        if(is_null($user)) {
+            return response()->json(['success'=> true, 'message' => 'no record found'], 403);
+        }
+        $user->verified = 1;
+        $user->save();
+        return response()->json(['success' => true], 200);
+    } 
 
     /**
      * Remove the specified resource from storage.
@@ -157,8 +213,12 @@ class UserAccountController extends ApiController
      * @param User $user
      * @return JsonResponse
      */
-    public function destroy(User $user)
+    public function destroy($userId)
     {
+        $user = User::find($userId);
+        if(is_null($user)) {
+            return response()->json(['success' => false, 'message' => "no record found"], 403); 
+        }
         $inspect = Gate::inspect('delete', $user);
         if($inspect->denied()) {
             return response()->json(['success' => false, 'errors' =>$inspect->message()], 403);
